TODO:实时数据数据库。反馈的bug大部分解决

+++ DeviceStatus/DataOverview/dataoverview.h
@@ -44,7 +44,7 @@ public:
         quint16 sys_single_hour;
         quint16 sys_single_min;
         quint16 sys_single_sec;
-        quint32 sys_total_hour;
+        quint16 sys_total_hour;
         quint16 sys_total_min;
         quint16 sys_total_sec;
     }Q_PACKED;
